+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.3.4:
+
+- New `allowedScriptHostnames` option, it enables you to specify which hostnames are allowed in a script tag
+- New `allowedScriptDomains` option, it enables you to specify which domains are allowed in a script tag
+
 ## 2.3.3 (2021-03-19):
 - Security fix: `allowedSchemes` and related options did not properly block schemes containing a hyphen, plus sign, period or digit, such as `ms-calculator:`. Thanks to Lukas Euler for pointing out the issue.
 - Added a security note about the known risks associated with using the `parser` option, especially `decodeEntities: false`. See the documentation.